modifies header test and deletes login test jsx
