+++ src/Components/Header/Header.test.js
@@ -1,23 +1,16 @@
 /* eslint-disable jest/no-identical-title */
 import { render, screen } from '@testing-library/react';
-import { MemoryRouter } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from '../../store/index';
 
 import Header from './Header';
 
 test('render link', () => {
-  render(<Header />, { wrapper: MemoryRouter });
+  render(<Provider store={store}><BrowserRouter><Header /></BrowserRouter></Provider>);
   // eslint-disable-next-line jest/valid-expect
   expect(screen.queryAllByText('navigation'));
-});
-test('renders text', () => {
-  render(<Header />, { wrapper: MemoryRouter });
-  expect(screen.getByText(/Home/i)).toBeInTheDocument();
-});
-test('renders text', () => {
-  render(<Header />, { wrapper: MemoryRouter });
-  expect(screen.getByText(/Users/i)).toBeInTheDocument();
-});
-test('renders text', () => {
-  render(<Header />, { wrapper: MemoryRouter });
-  expect(screen.getByText(/Doctors/i)).toBeInTheDocument();
+  expect(screen.getByText(/Inicio/i)).toBeInTheDocument();
+  expect(screen.getByText(/Usuarios/i)).toBeInTheDocument();
+  expect(screen.getByText(/Doctores/i)).toBeInTheDocument();
 });
